+++ project-scripts/gen_docs_from_playgrounds.sh
@@ -1,0 +1,15 @@
+#!/bin/sh
+
+# Description:
+#   Generates markdown documentation from nef playgrounds.
+#
+# Required environment variables:
+#   SRCROOT - path to project folder.
+#
+
+PLAYGROUNDS="${SRCROOT}/TIFoundationUtils/TIFoundationUtils.app"
+
+for playground_path in ${PLAYGROUNDS}; do
+    nef compile --project ${playground_path}
+    nef markdown --project ${playground_path} --output ../docs
+done